Bring back deep macro evaluation in tooltips
For example, hovering over BAR in:
> #define FOO 42
> #define BAR FOO + 56
> BAR

Will show a tooltip with "42 + 56" instead of "FOO + 56" as it would before.

The implementation is not perfect, it (ab)uses the IElementType like the old implementation did.
However making it the clean way would require implementing remapping system which allows using text.
And that could take some time.